+++ packages/desktop-client/src/components/budget/BudgetPageHeader.tsx
@@ -1,8 +1,6 @@
 // @ts-strict-ignore
 import React, { type ComponentProps, memo } from 'react';
 
-import * as monthUtils from 'loot-core/src/shared/months';
-
 import { View } from '../common/View';
 
 import { MonthPicker } from './MonthPicker';
@@ -17,18 +15,6 @@ type BudgetPageHeaderProps = {
 
 export const BudgetPageHeader = memo<BudgetPageHeaderProps>(
   ({ startMonth, onMonthSelect, numMonths, monthBounds }) => {
-    function getValidMonth(month) {
-      const start = monthBounds.start;
-      const end = monthUtils.subMonths(monthBounds.end, numMonths - 1);
-
-      if (month < start) {
-        return start;
-      } else if (month > end) {
-        return end;
-      }
-      return month;
-    }
-
     return (
       <View style={{ marginLeft: 200 + 5, flexShrink: 0 }}>
         <View style={{ marginRight: 5 + getScrollbarWidth() }}>
@@ -37,7 +23,7 @@ export const BudgetPageHeader = memo<BudgetPageHeaderProps>(
             numDisplayed={numMonths}
             monthBounds={monthBounds}
             style={{ paddingTop: 5 }}
-            onSelect={month => onMonthSelect(getValidMonth(month))}
+            onSelect={month => onMonthSelect(month)}
           />
         </View>
       </View>
